requirements.txt
fastapi[all]
motor
beanie
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings